More flexibility with Tessellation object in squares and adaptive_squares
I suggest to get rid of the "drop" statement and replace it with the selection of the columns to keep instead. The reason is that a Tessellation object created using a user-defined GeoDataFrame does likely not have the columns "osm_id" and "children_id", so they can't be dropped and the code results in an error.